get Products automapping
fast add

+++ src/Application/TodoItems/Commands/DeleteTodoItem/DeleteTodoItemCommand.cs
@@ -6,7 +6,7 @@ using MediatR;
 
 namespace CleanArchitecture.Application.TodoItems.Commands.DeleteTodoItem;
 
-public record DeleteTodoItemCommand(int Id) : IRequest;
+public record DeleteTodoItemCommand(long Id) : IRequest;
 
 public class DeleteTodoItemCommandHandler : IRequestHandler<DeleteTodoItemCommand>
 {
